requirements.txt
flask
pymongo
flask-cors
requests
pyjwt
werkzeug